Added lane parameters access to state machine and repaired output

+++ TUeLDT/UnitConversion.cpp
@@ -17,9 +17,9 @@ float UnitConversion::getDistance(int Y_ICCS, const Camera& CAM){
     int dy = CAM.RES_VH(0) * (CAM.FOV_HORIZON / CAM.FOV_VH(0) - 0.5 ); // Calculate pixels added/removed due to pitch angle.
     //thetaY = CAM.FOV_HORIZON * ((Y_ICCS - CAM.RES_VH(0))/CAM.RES_HORIZON + 1);
 
-    thetaY = CAM.FOV_HORIZON * ((Y_ICCS + dy) / CAM.RES_HORIZON);
+    thetaY = CAM.FOV_HORIZON * (((float)Y_ICCS + (float)dy) / (float)CAM.RES_HORIZON);
     float W = ty * 100 / tan(thetaY * M_PI / 180);
     //std::cout << "\t" << CAM.FOV_HORIZON << "\t" << thetaY << "\t" << Y_ICCS << "\t" << W << "\t" << CAM.RES_VH(0) <<endl;
-    //std::cout<<thetaY << "\t";
+    //std::cout<< "theyaY: " << thetaY << "\tdy: " << dy << "\tRes_Horizon: "<< CAM.RES_HORIZON << endl;
     return W;
 }
